Correct handling of IPv4 start nodes
Previously, the reader assumed that the IPv4 start would always be the
96th node. This is incorrect. The format does not specify the location
of the nodes in the database besides node 0. It is an artifact of the
current implementation of the MMDB writer that the 96th node is
generally the IPv4 start.

Also, a database may not have an IPv4 subtree at all, as in the case of
`MaxMind-DB-no-ipv4-search-tree.mmdb`, where there is one record that
cover all of the IPv4 space and it is higher up in the tree.